requirements.txt
Django==4.2.0
Pillow==11.1.0
django-redis==5.4.0
djangorestframework==3.15.2